fix: use typed data for intent signing
