+++ src/execution/utils.ts
@@ -7,11 +7,13 @@ import {
   type HashTypedDataParameters,
   type Hex,
   hashMessage,
+  keccak256,
   type PublicClient,
   publicActions,
   type SignableMessage,
   type SignedAuthorization,
   type SignedAuthorizationList,
+  slice,
   type TypedData,
   toHex,
   zeroAddress,
@@ -43,11 +45,11 @@ import {
   getWebAuthnValidator,
 } from '../modules/validators/core'
 import {
-  getIntentOpHash,
   getOrchestrator,
   type IntentInput,
   type IntentOp,
   type IntentRoute,
+  type SettlementSystem,
   type SignedIntentOp,
   type SupportedChain,
 } from '../orchestrator'
@@ -90,7 +92,6 @@ type TransactionResult =
 
 interface IntentData {
   type: 'intent'
-  hash: Hex
   intentRoute: IntentRoute
 }
 
@@ -177,7 +178,12 @@ async function signTransaction(
     // Smart sessions require a UserOp flow
     signature = await signUserOp(config, chain, signers, userOp)
   } else {
-    signature = await signIntent(config, targetChain, data.hash, signers)
+    signature = await signIntent(
+      config,
+      targetChain,
+      data.intentRoute.intentOp,
+      signers,
+    )
   }
 
   return {
@@ -455,19 +461,17 @@ async function prepareTransactionAsIntent(
     config.rhinestoneApiKey,
   )
   const intentRoute = await orchestrator.getIntentRoute(metaIntent)
-  const intentHash = getIntentOpHash(intentRoute.intentOp)
 
   return {
     type: 'intent',
     intentRoute,
-    hash: intentHash,
   } as IntentData
 }
 
 async function signIntent(
   config: RhinestoneAccountConfig,
   targetChainId: number,
-  intentHash: Hex,
+  intentOp: IntentOp,
   signers?: SignerSet,
 ) {
   const validator = getValidator(config, signers)
@@ -477,7 +481,11 @@ async function signIntent(
   const ownerValidator = getOwnerValidator(config)
   const isRoot = validator.address === ownerValidator.address
 
-  const signature = await getPackedSignature(
+  const notarizedChainElement = intentOp.elements[0]
+  const settlementSystem =
+    notarizedChainElement.mandate.qualifier.settlementSystem
+  const claimProofer = getClaimProofer(settlementSystem)
+  const signature = await getTypedDataPackedSignature(
     config,
     signers,
     targetChainId,
@@ -485,9 +493,104 @@ async function signIntent(
       address: validator.address,
       isRoot,
     },
-    intentHash,
+    {
+      domain: {
+        name: 'The Compact',
+        version: '1',
+        chainId: BigInt(notarizedChainElement.chainId),
+        verifyingContract: '0xAbd3388A633758D0Bae01Efb885EF1e87BD519a6',
+      },
+      types: {
+        MultichainCompact: [
+          { name: 'sponsor', type: 'address' },
+          { name: 'nonce', type: 'uint256' },
+          { name: 'expires', type: 'uint256' },
+          { name: 'elements', type: 'Element[]' },
+        ],
+        Element: [
+          { name: 'arbiter', type: 'address' },
+          { name: 'chainId', type: 'uint256' },
+          { name: 'commitments', type: 'Lock[]' },
+          { name: 'mandate', type: 'Mandate' },
+        ],
+        Lock: [
+          { name: 'lockTag', type: 'bytes12' },
+          { name: 'token', type: 'address' },
+          { name: 'amount', type: 'uint256' },
+        ],
+        Mandate: [
+          { name: 'target', type: 'Target' },
+          { name: 'originOps', type: 'Op[]' },
+          { name: 'destOps', type: 'Op[]' },
+          { name: 'q', type: 'bytes32' },
+        ],
+        Target: [
+          { name: 'recipient', type: 'address' },
+          { name: 'tokenOut', type: 'Token[]' },
+          { name: 'targetChain', type: 'uint256' },
+          { name: 'fillExpiry', type: 'uint256' },
+          { name: 'claimProofer', type: 'address' },
+        ],
+        Token: [
+          { name: 'token', type: 'address' },
+          { name: 'amount', type: 'uint256' },
+        ],
+        Op: [
+          { name: 'to', type: 'address' },
+          { name: 'value', type: 'uint256' },
+          { name: 'data', type: 'bytes' },
+        ],
+      },
+      primaryType: 'MultichainCompact',
+      message: {
+        sponsor: intentOp.sponsor,
+        nonce: BigInt(intentOp.nonce),
+        expires: BigInt(intentOp.expires),
+        elements: intentOp.elements.map((element) => ({
+          arbiter: element.arbiter,
+          chainId: BigInt(element.chainId),
+          commitments: element.idsAndAmounts.map((token) => ({
+            lockTag: slice(toHex(BigInt(token[0])), 0, 12),
+            token: slice(toHex(BigInt(token[0])), 12, 32),
+            amount: BigInt(token[1]),
+          })),
+          mandate: {
+            target: {
+              recipient: element.mandate.recipient,
+              tokenOut: element.mandate.tokenOut.map((token) => ({
+                token: slice(toHex(BigInt(token[0])), 12, 32),
+                amount: BigInt(token[1]),
+              })),
+              targetChain: BigInt(element.mandate.destinationChainId),
+              fillExpiry: BigInt(element.mandate.fillDeadline),
+              claimProofer: claimProofer,
+            },
+            originOps: element.mandate.preClaimOps.map((op) => ({
+              to: op.to,
+              value: BigInt(op.value),
+              data: op.data,
+            })),
+            destOps: element.mandate.destinationOps.map((op) => ({
+              to: op.to,
+              value: BigInt(op.value),
+              data: op.data,
+            })),
+            q: keccak256(element.mandate.qualifier?.encodedVal ?? '0x'),
+          },
+        })),
+      },
+    },
   )
   return signature
+}
+
+function getClaimProofer(settlementSystem: SettlementSystem): Address {
+  switch (settlementSystem) {
+    case 'ACROSS':
+      return '0x1636b30481Db91Bbc5818e65d3962838BdCd5569'
+    case 'SAME_CHAIN':
+      return zeroAddress
+  }
 }
 
 async function signUserOp(
